third_party/WebKit/LayoutTests/http/tests/htmlimports/block-cookies.html

<!DOCTYPE html>
<html>
<head>
<link id="target" rel="import" href="resources/cookie.cgi">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
shouldBe("target.import.querySelector('h1').innerHTML", "'Cookie Sent!'");
shouldBe("document.cookie.indexOf('htmlimport=hello')", "-1");
</script>
</body>
</html>
