third_party/WebKit/LayoutTests/http/tests/htmlimports/block-cookies-expected.txt

PASS target.import.querySelector('h1').innerHTML is 'Cookie Sent!'
PASS document.cookie.indexOf('htmlimport=hello') is -1
PASS successfullyParsed is true

TEST COMPLETE
